@@ -1,6 +1,14 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.0.10'
+gem 'rails', '3.1.0'
+
+#gem 'enju_amazon', :git => 'git://github.com/nabeta/enju_amazon.git'
+gem 'enju_barcode', :git => 'git://github.com/nabeta/enju_barcode.git'
+#gem 'enju_calil', :git => 'git://github.com/nabeta/enju_calil.git'
+#gem 'enju_ndl', :git => 'git://github.com/nabeta/enju_ndl.git'
+gem 'enju_nii', :git => 'git://github.com/nabeta/enju_nii.git'
+gem 'enju_oai', :git => 'git://github.com/nabeta/enju_oai.git'
+#gem 'enju_scribd', :git => 'git://github.com/nabeta/enju_scribd.git'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
@@ -34,55 +42,57 @@ end
 
 gem 'fastercsv' if RUBY_VERSION < '1.9'
 
-gem 'will_paginate', '~>3.0'
-gem 'exception_notification', '~>2.5.2'
+gem 'will_paginate', '~> 3.0'
+gem 'exception_notification', '~> 2.5.2'
 gem 'delayed_job', '>=2.1.4'
 gem 'state_machine'
-gem 'sunspot_rails', '>=1.2.1'
-gem 'friendly_id', '~>3.3'
-gem 'inherited_resources', '~>1.3'
+gem 'sunspot_rails', '~>1.3.0.rc4'
+gem 'sunspot_solr', '~>1.3.0.rc4'
+gem 'progress_bar'
+gem "friendly_id", "~> 4.0.0.beta12"
+gem 'inherited_resources', '~> 1.3'
 gem 'has_scope'
 gem 'nokogiri'
 gem 'marc'
-gem 'strongbox', '>=0.4.8'
-gem 'acts-as-taggable-on', '~>2.1'
+gem 'strongbox', '>= 0.4.8'
+gem 'acts-as-taggable-on', '~> 2.1'
 gem 'dalli', '~>1.1'
-gem 'sitemap_generator', '~>2.1'
+gem 'sitemap_generator', '~> 2.1'
 gem 'ri_cal'
 gem 'file_wrapper'
-gem 'paper_trail', '~>2.3'
-gem 'recurrence'
-gem 'prism'
-gem 'money'
-gem 'RedCloth', '>=4.2.8'
+gem 'paper_trail', '~> 2.3'
+#gem 'recurrence'
+#gem 'prism'
+#gem 'money'
+gem 'RedCloth', '>= 4.2.8'
 gem 'isbn-tools', :git => 'git://github.com/nabeta/isbn-tools.git', :require => 'isbn/tools'
 gem 'attribute_normalizer'
 gem 'configatron'
 gem 'extractcontent'
-gem 'cancan', '>=1.6.5'
-gem 'scribd_fu'
-gem 'devise', '~>1.4'
-gem 'omniauth', '>=0.2.6'
-gem 'paperclip', '~>2.4'
-gem 'whenever', '~>0.6', :require => false
-gem 'amazon-ecs', '>=2.2.0', :require => 'amazon/ecs'
-gem 'aws-s3', :require => 'aws/s3'
+gem 'cancan', '>= 1.6.5'
+gem 'devise', '~> 1.4.6'
+gem 'omniauth', '>= 0.2.6'
+gem 'paperclip', '~> 2.4'
+gem 'whenever', '~> 0.6', :require => false
+#gem 'amazon-ecs', '>= 2.2.0', :require => 'amazon/ecs'
+#gem 'aws-s3', :require => 'aws/s3'
 gem 'astrails-safe'
 gem 'dynamic_form'
 gem 'sanitize'
-gem 'barby', '~>0.5'
-gem 'rqrcode'
-gem 'jpmobile', '~>1.0'
+#gem 'barby', '~> 0.5'
+#gem 'rqrcode'
+gem 'jpmobile', '~> 2.0'
 #gem 'geokit'
 gem 'geocoder'
 gem 'acts_as_list', :git => 'https://github.com/swanandp/acts_as_list'
 gem 'library_stdnums'
 gem 'client_side_validations'
-gem 'simple_form', '~>1.5'
+gem 'simple_form', '~> 1.5'
 gem 'validates_timeliness'
 gem 'rack-protection'
+gem 'rails_autolink'
 
-gem 'oink', '>=0.9.2'
+gem 'oink', '>= 0.9.2'
 
 group :development do
   gem 'parallel_tests'
@@ -92,9 +102,9 @@ end
 group :development, :test do
   gem 'rspec-rails'
   gem 'guard-rspec'
-  gem 'factory_girl_rails', '~>1.2'
+  gem 'factory_girl_rails', '~> 1.2'
   gem 'spork', '~> 0.9.0.rc9'
-  gem 'metric_fu', '~>2.1'
+  gem 'metric_fu', '~> 2.1'
   gem 'timecop'
 end
 
