@@ -12,8 +12,7 @@ class ManifestationsController < ApplicationController
   after_filter :solr_commit, :only => [:create, :update, :destroy]
   after_filter :convert_charset, :only => :index
   cache_sweeper :manifestation_sweeper, :only => [:create, :update, :destroy]
-  #include WorldcatController
-  include OaiController
+  include EnjuOai::OaiController
 
   # GET /manifestations
   # GET /manifestations.xml
@@ -551,7 +550,7 @@ class ManifestationsController < ApplicationController
     @roles = Role.all
     @languages = Language.all_cache
     @frequencies = Frequency.all
-    @nii_types = NiiType.all
+    @nii_types = NiiType.all if defined?(NiiType)
   end
 
   def save_search_history(query, offset = 0, total = 0, user = nil)
