@@ -7,9 +7,9 @@ class PictureFilesController < ApplicationController
   # GET /picture_files.xml
   def index
     if @attachable
-      @picture_files = @attachable.picture_files.attached.paginate(:page => params[:page])
+      @picture_files = @attachable.picture_files.attached.page(params[:page])
     else
-      @picture_files = PictureFile.attached.paginate(:page => params[:page])
+      @picture_files = PictureFile.attached.page(params[:page])
     end
 
     respond_to do |format|
