@@ -54,7 +54,8 @@ class User < ActiveRecord::Base
   after_save :index_patron
   after_destroy :index_patron
 
-  has_friendly_id :username
+  extend FriendlyId
+  friendly_id :username
   #acts_as_tagger
   has_paper_trail
   normalize_attributes :username, :user_number #, :email
