@@ -8,7 +8,7 @@ describe PictureFilesController do
   describe "GET index" do
     describe "When logged in as Administrator" do
       before(:each) do
-        sign_in Factory(:admin)
+        sign_in FactoryGirl.create(:admin)
       end
 
       it "assigns all picture_files as @picture_files" do
@@ -19,7 +19,7 @@ describe PictureFilesController do
 
     describe "When logged in as Librarian" do
       before(:each) do
-        sign_in Factory(:librarian)
+        sign_in FactoryGirl.create(:librarian)
       end
 
       it "assigns all picture_files as @picture_files" do
@@ -30,7 +30,7 @@ describe PictureFilesController do
 
     describe "When logged in as User" do
       before(:each) do
-        sign_in Factory(:user)
+        sign_in FactoryGirl.create(:user)
       end
 
       it "assigns all picture_files as @picture_files" do
@@ -50,7 +50,7 @@ describe PictureFilesController do
   describe "GET show" do
     describe "When logged in as Administrator" do
       before(:each) do
-        sign_in Factory(:admin)
+        sign_in FactoryGirl.create(:admin)
       end
 
       it "assigns the requested picture_file as @picture_file" do
@@ -62,7 +62,7 @@ describe PictureFilesController do
 
     describe "When logged in as Librarian" do
       before(:each) do
-        sign_in Factory(:librarian)
+        sign_in FactoryGirl.create(:librarian)
       end
 
       it "assigns the requested picture_file as @picture_file" do
@@ -74,7 +74,7 @@ describe PictureFilesController do
 
     describe "When logged in as User" do
       before(:each) do
-        sign_in Factory(:user)
+        sign_in FactoryGirl.create(:user)
       end
 
       it "assigns the requested picture_file as @picture_file" do
@@ -96,7 +96,7 @@ describe PictureFilesController do
   describe "GET new" do
     describe "When logged in as Administrator" do
       before(:each) do
-        sign_in Factory(:admin)
+        sign_in FactoryGirl.create(:admin)
       end
 
       it "assigns the requested picture_file as @picture_file" do
@@ -108,7 +108,7 @@ describe PictureFilesController do
 
     describe "When logged in as Librarian" do
       before(:each) do
-        sign_in Factory(:librarian)
+        sign_in FactoryGirl.create(:librarian)
       end
 
       it "assigns the requested picture_file as @picture_file" do
@@ -120,7 +120,7 @@ describe PictureFilesController do
 
     describe "When logged in as User" do
       before(:each) do
-        sign_in Factory(:user)
+        sign_in FactoryGirl.create(:user)
       end
 
       it "should not assign the requested picture_file as @picture_file" do
@@ -142,7 +142,7 @@ describe PictureFilesController do
   describe "GET edit" do
     describe "When logged in as Administrator" do
       before(:each) do
-        sign_in Factory(:admin)
+        sign_in FactoryGirl.create(:admin)
       end
 
       it "assigns the requested picture_file as @picture_file" do
@@ -154,7 +154,7 @@ describe PictureFilesController do
 
     describe "When logged in as Librarian" do
       before(:each) do
-        sign_in Factory(:librarian)
+        sign_in FactoryGirl.create(:librarian)
       end
 
       it "assigns the requested picture_file as @picture_file" do
@@ -166,7 +166,7 @@ describe PictureFilesController do
 
     describe "When logged in as User" do
       before(:each) do
-        sign_in Factory(:user)
+        sign_in FactoryGirl.create(:user)
       end
 
       it "assigns the requested picture_file as @picture_file" do
@@ -187,13 +187,13 @@ describe PictureFilesController do
 
   describe "POST create" do
     before(:each) do
-      @attrs = {:picture_attachable_type => 'Shelf', :picture_attachable_id => 1, :picture => File.new("#{Rails.root.to_s}/examples/spinner.gif")}
+      @attrs = {:picture_attachable_type => 'Shelf', :picture_attachable_id => 1, :picture => fixture_file_upload("#{Rails.root.to_s}/examples/spinner.gif")}
       @invalid_attrs = {:filename => ''}
     end
 
     describe "When logged in as Administrator" do
       before(:each) do
-        sign_in Factory(:admin)
+        sign_in FactoryGirl.create(:admin)
       end
 
       describe "with valid params" do
@@ -223,7 +223,7 @@ describe PictureFilesController do
 
     describe "When logged in as Librarian" do
       before(:each) do
-        sign_in Factory(:librarian)
+        sign_in FactoryGirl.create(:librarian)
       end
 
       describe "with valid params" do
@@ -253,7 +253,7 @@ describe PictureFilesController do
 
     describe "When logged in as User" do
       before(:each) do
-        sign_in Factory(:user)
+        sign_in FactoryGirl.create(:user)
       end
 
       describe "with valid params" do
@@ -317,7 +317,7 @@ describe PictureFilesController do
 
     describe "When logged in as Administrator" do
       before(:each) do
-        sign_in Factory(:admin)
+        sign_in FactoryGirl.create(:admin)
       end
 
       describe "with valid params" do
@@ -346,7 +346,7 @@ describe PictureFilesController do
 
     describe "When logged in as Librarian" do
       before(:each) do
-        sign_in Factory(:librarian)
+        sign_in FactoryGirl.create(:librarian)
       end
 
       describe "with valid params" do
@@ -376,7 +376,7 @@ describe PictureFilesController do
 
     describe "When logged in as User" do
       before(:each) do
-        sign_in Factory(:user)
+        sign_in FactoryGirl.create(:user)
       end
 
       describe "with valid params" do
@@ -427,7 +427,7 @@ describe PictureFilesController do
 
     describe "When logged in as Administrator" do
       before(:each) do
-        sign_in Factory(:admin)
+        sign_in FactoryGirl.create(:admin)
       end
 
       it "destroys the requested picture_file" do
@@ -442,7 +442,7 @@ describe PictureFilesController do
 
     describe "When logged in as Librarian" do
       before(:each) do
-        sign_in Factory(:librarian)
+        sign_in FactoryGirl.create(:librarian)
       end
 
       it "destroys the requested picture_file" do
@@ -457,7 +457,7 @@ describe PictureFilesController do
 
     describe "When logged in as User" do
       before(:each) do
-        sign_in Factory(:user)
+        sign_in FactoryGirl.create(:user)
       end
 
       it "destroys the requested picture_file" do
